feat: add `NuxtAppCompat` interface
